Move genaccounts code to be reused
