+++ client/genaccounts/main.go
@@ -1,4 +1,4 @@
-package main
+package genaccounts
 
 import (
 	"errors"
